+++ app/src/test/java/com/example/mankomania/logik/NoteTypesTest.java
@@ -1,12 +1,8 @@
-package com.example.mankomania.LogikTest;
+package com.example.mankomania.logik;
 
 import static org.junit.Assert.assertEquals;
 
-import com.example.mankomania.logik.NoteTypes;
-
 import org.junit.Test;
-
-import java.util.Arrays;
 
 public class NoteTypesTest {
