+++ app/src/test/java/com/example/mankomania/logik/PlayerTest.java
@@ -1,15 +1,9 @@
-package com.example.mankomania.LogikTest;
+package com.example.mankomania.logik;
 
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertTrue;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-import com.example.mankomania.logik.Color;
-import com.example.mankomania.logik.NoteTypes;
-import com.example.mankomania.logik.Player;
-import com.example.mankomania.logik.StockTypes;
-
-import org.junit.Assert;
 import org.junit.Test;
 
 public class PlayerTest {
