+++ app/src/test/java/com/example/mankomania/logik/StockMarketTest.java
@@ -1,15 +1,11 @@
-package com.example.mankomania.LogikTest;
+package com.example.mankomania.logik;
 
 import static org.junit.Assert.assertTrue;
 import static org.junit.Assert.fail;
 import static org.junit.jupiter.api.Assertions.assertNotEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 
-import com.example.mankomania.logik.StockMarket;
-import com.example.mankomania.logik.StockTypes;
-
 import org.junit.Test;
-import org.junit.jupiter.api.BeforeEach;
 
 public class StockMarketTest {
     private StockMarket stockMarket = new StockMarket();
